Modificando base de datos

+++ backend/modelos/modelos.py
@@ -28,10 +28,10 @@ class Producto(db.Model):
     nombre_prod = db.Column(db.String(100))
     medida_prod = db.Column(db.Integer)
     unidad_medida = db.Column(db.String(50))
-    precio_bruto_prod = db.Column(db.Numeric(19, 0))
-    iva_prod = db.Column(db.Numeric(3, 2))
-    porcentaje_ganancia = db.Column(db.Numeric(3, 2))
-    precio_neto_unidad_prod = db.Column(db.Numeric(19, 2))
+    precio_bruto_prod = db.Column(db.Integer)
+    iva_prod = db.Column(db.Float)
+    porcentaje_ganancia = db.Column(db.Float)
+    precio_neto_unidad_prod = db.Column(db.Float)
     unidades_totales = db.Column(db.Integer)
     marca_prod = db.Column(db.String(60))
     estado_prod = db.Column(db.String(50))
@@ -65,20 +65,20 @@ class Usuario(db.Model):
 
 class Factura(db.Model):
     id = db.Column(db.Integer, primary_key = True)
-    impuestos_fact = db.Column(db.Numeric(5, 2))
+    impuestos_fact = db.Column(db.Float)
     fecha_generacion_fact = db.Column(db.DateTime)
 
 class Venta(db.Model):
     id = db.Column(db.Integer, primary_key = True)
     fecha_venta = db.Column(db.DateTime)
-    total_venta = db.Column(db.Numeric(19, 0))
+    total_venta = db.Column(db.Integer)
     forma_pago = db.Column(db.String(50))
     usuario = db.Column(db.Integer, db.ForeignKey("usuario.id"))
 
 class Detalle_Venta(db.Model):
     id = db.Column(db.Integer, primary_key = True)
     cantidad = db.Column(db.Integer)
-    precio_unidad = db.Column(db.Numeric(19, 0))
+    precio_unidad = db.Column(db.Integer)
     venta = db.Column(db.Integer, db.ForeignKey("venta.id"))
     factura = db.Column(db.Integer, db.ForeignKey("factura.id"))
     producto = db.Column(db.Integer, db.ForeignKey("producto.id"))
